moved weights file logic out into weights.c (need to check whether the header files make sense etc but it's compiling)
