+++ src/plm.c
@@ -19,9 +19,9 @@
 #endif
 
 #include "include/twister.h"
-
 #include "include/plm.h"
 #include "include/inference.h"
+#include "include/weights.h"
 
 /* Usage pattern */
 const char *usage =
@@ -56,6 +56,7 @@ const char *usage =
 "      -g  --gapignore                  Model sequence likelihoods only by coding, non-gapped portions\n"
 "      -m  --maxiter                    Maximum number of iterations\n"
 "      -n  --ncores    [<number>|max]   Maximum number of threads to use in OpenMP\n"
+"      -w  --weights   weightsfile      Load sequence weights from file\n"
 "      -h  --help                       Usage\n\n";
 
 /* Internal functions to MSARead */
@@ -162,7 +163,8 @@ int main(int argc, char **argv) {
         } else if ((arg < argc-1) && strcmp(argv[arg], "--fast") == 0) {
             options->sgd = 1;
             options->fastWeights = 100;
-        } else if ((arg < argc-1) && strcmp(argv[arg], "--weights_file") == 0) {
+        } else if ((arg < argc-1) && (strcmp(argv[arg], "--weights_file") == 0
+                    || strcmp(argv[arg], "-w") == 0)) {
             weightsFile = argv[++arg];
         } else if ((arg < argc-1) && (strcmp(argv[arg], "--ncores") == 0
                     || strcmp(argv[arg], "-n") == 0)) {
@@ -205,7 +207,7 @@ int main(int argc, char **argv) {
     }
     alignFile = argv[argc - 1];
 
-    /* Read multiple seqence alignment */
+    /* Read multiple sequence alignment */
     alignment_t *ali = MSARead(alignFile, options);
 
     if (weightsFile != NULL) {
@@ -216,7 +218,7 @@ int main(int argc, char **argv) {
         MSAReweightSequences(ali, options);
     }
 
-    /* Compute sitwise and pairwise marginal distributions */
+    /* Compute sitewise and pairwise marginal distributions */
     MSACountMarginals(ali, options);
 
     /* Infer model parameters */
@@ -544,309 +546,6 @@ letter_t MSAReadCode(char c, char *alphabet, int nCodes) {
     /* Encode out-of-alphabet characters by [nCodes] */
     if (i > 0 && toupper(c) != alphabet[i]) i = nCodes;
     return i;
-}
-
-void MSAReweightSequences(alignment_t *ali, options_t *options) {
-    /* Reweight seqeuences by their inverse neighborhood size. Each sequence's
-       weight is the inverse of the number of neighboring sequences with less
-       than THETA percent divergence
-    */
-    for (int i = 0; i < ali->nSeqs; i++) ali->weights[i] = 1.0;
-
-    /* Only apply reweighting if theta is on [0,1] */
-    if (options->theta >= 0 && options->theta <= 1) {
-        /* The neighborhood size of each sequence is the number of sequences 
-           in the alignment within theta percent divergence */
-
-        if (options->fastWeights > 0 && options->fastWeights < ali->nSeqs) {
-            /* Cluster the sequences with k-consensus */
-            int nClusters = options->fastWeights;
-            int nIterations = 10;
-            int nSeqs = ali->nSeqs;
-            int nCodes = ali->nCodes;
-            int nSites = ali->nSites;
-
-            #define COUNTS(i,j,a) counts[i * nSites * nCodes + j * nCodes + a]
-            #define CONSENSUS(i,j) consensus[i * nSites + j]
-            #define ALI(i,j) aliPermute[i * nSites + j]
-
-            /* Pick initial clusters with Reservoir sampling */
-            int *clusters = (int *) malloc(nClusters * sizeof(int));
-            letter_t *consensus = (letter_t *) malloc(nClusters * nSites * sizeof(letter_t));
-            for (int i = 0; i < nClusters; i++) clusters[i] = i;
-            for (int i = nClusters; i < nSeqs; i++) {
-                int ix = genrand_int32() % (i);
-                if (ix < nClusters) clusters[ix] = i;
-            }
-            for (int i = 0; i < nClusters; i++)
-                for (int j = 0; j < nSites; j++)
-                    CONSENSUS(i,j) = seq(clusters[i], j);
-            free(clusters);
-
-            /* EM steps */
-            int *assignment = (int *) malloc(nSeqs * sizeof(int));
-            int *counts = (int *) malloc(nClusters * nSites * nCodes * sizeof(int));
-            int *radii = (int *) malloc(nClusters * sizeof(int));
-            for (int i = 0; i < nSeqs; i++) assignment[i] = 0;
-            fprintf(stderr, "Clustering");
-            for (int t = 0; t < nIterations; t++) {
-                fprintf(stderr, ".");
-                /* Step 1. Update the assignments */
-                for (int i = 0; i < nClusters; i++) radii[i] = 0;
-                #pragma omp parallel for
-                for (int s = 0; s < nSeqs; s++) {
-                    int ixOld = assignment[s];
-                    /* Current distance to current assignment */
-                    numeric_t distance = 0;
-                    for (int j = 0; j < nSites; j++)
-                        distance += (CONSENSUS(ixOld, j) != seq(s, j));
-                    /* Find closest */
-                    int ixNew = ixOld;
-                    for (int i = 0; i < nClusters; i++) {
-                        numeric_t distanceI = 0;
-                        for (int j = 0; j < nSites; j++)
-                            distanceI += (CONSENSUS(i, j) != seq(s, j));
-                        if (distanceI < distance) {
-                            ixNew = i;
-                            distance = distanceI;
-                        }
-                    }
-                    if (ixNew != ixOld) assignment[s] = ixNew;
-                    if (radii[ixNew] < distance) radii[ixNew] = distance;
-                }
-                /* --------------------------_DEBUG_--------------------------*/
-                // for (int s = 0; s < nClusters; s++) {
-                //     int size = 0;
-                //     for (int i = 0; i < nSeqs; i++) size += (assignment[i] == s);
-                //     fprintf(stderr, ">Cluster %d, %d members, radius %d\n", s, size, radii[s]);
-                //     for (int i = 0; i < ali->nSites; i++)
-                //         if (CONSENSUS(s,i) >= 0) {
-                //             fprintf(stderr, "%c", ali->alphabet[CONSENSUS(s,i)]);
-                //         } else {
-                //             fprintf(stderr, " ");
-                //         }
-                //     fprintf(stderr, "\n");
-                // }
-                /* --------------------------^DEBUG^--------------------------*/
-
-                /* Step 2. Update the consensus sequences */
-                /* Update the counts */
-                if (t < nIterations - 1) {
-                    for (int i = 0; i < nClusters * nSites * nCodes; i++)
-                        counts[i] = 0;
-                    for (int s = 0; s < nSeqs; s++)
-                        for (int j = 0; j < nSites; j++)
-                            COUNTS(assignment[s], j, seq(s, j)) += 1;
-                    #pragma omp parallel for
-                    for (int i = 0; i < nClusters; i++)
-                        for (int j = 0; j < nSites; j++) {
-                            int topCode = 0;
-                            int topCounts = COUNTS(i, j, 0);
-                            for (int b = 1; b < nCodes; b++)
-                                if (COUNTS(i, j, b) > topCounts) {
-                                    topCode = b;
-                                    topCounts = COUNTS(i, j, b);
-                                }
-                            CONSENSUS(i ,j) = topCode;
-                        }
-                }
-            }
-            fprintf(stderr, "\n");
-
-            /* Profile-profile distances */
-            numeric_t *clusterID = (numeric_t *) malloc(nClusters * nClusters * sizeof(numeric_t));
-            for (int i = 0; i < nClusters * nClusters; i++) clusterID[i] = 0;
-            #pragma omp parallel for
-            for (int pi = 0; pi < nClusters; pi++)
-                for (int pj = 0; pj < nClusters; pj++)
-                    for (int j = 0; j < nSites; j++)
-                        clusterID[pi + pj * nClusters] += 
-                            (CONSENSUS(pi,j) == CONSENSUS(pj,j));
-
-            free(consensus);
-            free(counts);
-
-            /* Permute alignment */
-            int *clusterSizes = (int *) malloc(nClusters * sizeof(int));
-            int *clusterStart = (int *) malloc(nClusters * sizeof(int));
-            int *clusterEnd = (int *) malloc(nClusters * sizeof(int));
-            int *permuteMap = (int *) malloc(nSeqs * sizeof(int));
-            numeric_t *weightsP = (numeric_t *) malloc(nSeqs * sizeof(numeric_t));
-            letter_t *aliPermute = (letter_t *) malloc(nSeqs * nSites * sizeof(letter_t));
-            for (int i = 0; i < nClusters; i++) clusterSizes[i] = 0;
-            for (int s = 0; s < ali->nSeqs; s++)
-                clusterSizes[assignment[s]] += 1;
-            int ix = 0;
-            for (int i = 0; i < nClusters; i++) {
-                clusterStart[i] = ix;
-                ix += clusterSizes[i];
-                clusterEnd[i] = ix;
-            }
-            ix = 0;
-            for (int i = 0; i < nClusters; i++)
-                for (int s = 0; s < ali->nSeqs; s++)
-                    if (assignment[s] == i) {
-                        for (int j = 0; j < nSites; j++)
-                            ALI(ix,j) = seq(s,j);
-                        permuteMap[ix] = s;
-                        ix++;
-                    }
-
-            /* ----------------------------_DEBUG_----------------------------*/
-            // for (int s = 0; s < nSeqs; s++) {
-            //     fprintf(stdout, ">Seq %d\n", s);
-            //     for (int i = 0; i < ali->nSites; i++)
-            //             fprintf(stdout, "%c", ali->alphabet[ALI(s,i)]);
-            //     fprintf(stdout, "\n");
-            // }
-            /* ----------------------------^DEBUG^----------------------------*/
-
-            /* Sequence weights */
-            numeric_t cutoff = (numeric_t) ((1 - options->theta) * ali->nSites);
-            for (int s = 0; s < nSeqs; s++) weightsP[s] = 1;
-            #pragma omp parallel for
-            for (int ci = 0; ci < nClusters; ci++)
-                for (int cj = 0; cj < nClusters; cj++)
-                    if (clusterID[ci * nClusters + cj] >= 0.9 * cutoff)
-                        for (int s = clusterStart[ci]; s < clusterEnd[ci]; s++)
-                            for (int t = clusterStart[cj]; t < clusterEnd[cj]; t++)
-                                if (s != t) {
-                                    int id = 0;
-                                    for (int n = 0; n < ali->nSites; n++)
-                                        id += (ALI(s, n) == ALI(t, n));
-                                    if (id >= cutoff) weightsP[s] += 1.0;
-                                }
-            for (int s = 0; s < nSeqs; s++)
-                ali->weights[permuteMap[s]] = weightsP[s];
-
-            #undef COUNTS
-            #undef CONSENSUS
-            #undef ALI
-
-            free(clusterSizes);
-            free(clusterStart);
-            free(clusterEnd);
-            free(permuteMap);
-            free(weightsP);
-            free(radii);
-            free(aliPermute);
-        } else {
-            /* Deterministic sequence weights */
-            #if defined(_OPENMP)
-            /* Naive parallelization is faster ignoring symmetry */
-            #pragma omp parallel for
-            for (int s = 0; s < ali->nSeqs; s++)
-                for (int t = 0; t < ali->nSeqs; t++)
-                    if (s != t) {
-                        int id = 0;
-                        for (int n = 0; n < ali->nSites; n++)
-                            id += (seq(s, n) == seq(t, n));
-                        if (id >= ((1 - options->theta) * ali->nSites))
-                            ali->weights[s] += 1.0;
-                    }
-            #else
-            /* For a single core, take advantage of symmetry */
-            for (int s = 0; s < ali->nSeqs - 1; s++)
-                for (int t = s + 1; t < ali->nSeqs; t++) {
-                    int id = 0;
-                    for (int n = 0; n < ali->nSites; n++)
-                        id += (seq(s, n) == seq(t, n));
-                    if (id >= ((1 - options->theta) * ali->nSites)) {
-                        ali->weights[s] += 1.0;
-                        ali->weights[t] += 1.0;
-                    }
-                }
-            #endif
-        }
-
-        /* Reweight sequences by the inverse of the neighborhood size */
-        for (int i = 0; i < ali->nSeqs; i++)
-            ali->weights[i] = 1.0 / ali->weights[i];
-    }
-
-    /* Scale sets the effective number of samples per neighborhood */
-    for (int i = 0; i < ali->nSeqs; i++)
-            ali->weights[i] *= options->scale;
-
-    /* The effective number of sequences is then the sum of the weights */
-    ali->nEff = 0;
-    for (int i = 0; i < ali->nSeqs; i++) ali->nEff += ali->weights[i];
-
-    if (options->theta >= 0 && options->theta <= 1) {
-        fprintf(stderr, 
-            "Effective number of samples (to 1 decimal place): %.1f\t(%.0f%% identical neighborhood = %.3f samples)\n",
-            ali->nEff, 100 * (1 - options->theta), options->scale);
-    } else {
-        fprintf(stderr,
-            "Theta not between 0 and 1, no sequence reweighting applied (N = %.2f)\n",
-            ali->nEff);
-    }
-}
-
-int ValidateCustomWeightsFile(alignment_t *ali, options_t *options, char *weightsFile) {
-    /* Check that the weights file exists */
-    /* Remember to close file pointer before returning. */
-    FILE *fp = fopen(weightsFile, "r");
-    if (fp == NULL) {
-        fprintf(stderr, "Error: weights file %s does not exist\n", weightsFile);
-        fclose(fp);
-        return 1;
-    }
-
-    /* Count number of lines in file */
-    int nLines = 0;
-    int MAX_LINE_LENGTH = 1024;
-    char line[MAX_LINE_LENGTH];
-    while (fgets(line, MAX_LINE_LENGTH, fp) != NULL) nLines++;
-
-    if (nLines != ali->nSeqs) {
-        fprintf(stderr, "Error: weights file %s has %d lines, but alignment has %d sequences\n",
-            weightsFile, nLines, ali->nSeqs);
-        fclose(fp);
-        return 1;
-    }
-
-    fclose(fp);
-    return 0;
-}
-
-void ReadCustomWeightsFile(alignment_t *ali, options_t *options, char *weightsFile) {
-    /* Note: Not using options->scale (or options->theta) for now (assuming this is done in original weights calc).*/
-    /* Most of this is copied from MSAReweightSequences() */
-
-    int isValid = ValidateCustomWeightsFile(ali, options, weightsFile);
-    if (!isValid) {
-        fprintf(stderr, "Error: weights file %s is invalid\n", weightsFile);
-        exit(1);
-    }
-
-    /* Load weights (float array) into ali->weights and set ali->nEff */
-    FILE *fp = fopen(weightsFile, "r");
-//    if (fp == NULL) {
-//        fprintf(stderr, "Error: could not open weights file %s\n", weightsFile);
-//        exit(1);
-//    }
-    /* Reinitialize array just in case */
-    for (int i = 0; i < ali->nSeqs; i++) ali->weights[i] = 1.0;
-
-    /* Read weights, one float per line */
-    for (int i = 0; i < ali->nSeqs; i++) {
-        float w;
-        if (fscanf(fp, "%f", &w) != 1) {
-            fprintf(stderr, "Error reading weights file %s at position %d\n", weightsFile, i);
-            exit(1);
-        }
-        ali->weights[i] = w;
-    }
-    fclose(fp);
-
-    /* The effective number of sequences is then the sum of the weights */
-    ali->nEff = 0;
-    for (int i = 0; i < ali->nSeqs; i++) ali->nEff += ali->weights[i];
-
-    fprintf(stderr,
-            "Weights loaded successfully. Effective number of samples (to 1 decimal place): %.1f.\n",
-            ali->nEff);
 }
 
 void MSACountMarginals(alignment_t *ali, options_t *options) {
